Move do_* functions operating on files only to file_syscalls.c
